% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_ref_table.R
\name{label_ref_snp}
\alias{label_ref_snp}
\title{Label one allele (A/C/T/G) as reference and the other as other for use in determining numeric count genotypes.}
\usage{
label_ref_snp(gv_actg)
}
\arguments{
\item{gv_actg}{A genotype vector (coded as A/C/T/G) for a single SNP locus.}
}
\value{
a character vector of length 2 with names reference and other
}
\description{
Label one allele (A/C/T/G) as reference and the other as other for use in determining numeric count genotypes.
}
\examples{
label_ref_snp(c("TT", "CT", "CC", "TC"))
}

