% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_ref_table.R
\name{make_ref_table}
\alias{make_ref_table}
\title{Create a reference table for the conversion of A/C/T/G genotype vectors (for a single locus) to 0,1,2 allele count genotypes.}
\usage{
make_ref_table(gv_actg)
}
\arguments{
\item{gv_actg}{A genotype vector (coded as A/C/T/G) for a single SNP locus.}
}
\value{
a reference table with 1) inputted vector, 2) labeled alleles and 3) numeric count genotypes (0,1,2)
}
\description{
Create a reference table for the conversion of A/C/T/G genotype vectors (for a single locus) to 0,1,2 allele count genotypes.
}
\examples{
make_ref_table(c("TT", "CT", "CC", "TC"))
}

